Implemented Read System Call
Implemented the read system call callback function. This function
detects when a target source is being read from, and taints the read
buffer so that it may be later queried. Also implemented an argument
which turns on tainting at some instruction number, to make it faster to
debug the plugin.